Fmt.Internal.Generic

Documentation

>>> import Fmt

genericF :: (Generic a, GBuildable (Rep a)) => a -> Builder

Format an arbitrary value without requiring a Buildable instance:

>>> data Foo = Foo { x :: Bool, y :: [Int] } deriving Generic

>>> fmt (genericF (Foo True [1,2,3]))
Foo:
  x: True
  y: [1, 2, 3]

It works for non-record constructors too:

>>> data Bar = Bar Bool [Int] deriving Generic

>>> fmtLn (genericF (Bar True [1,2,3]))
<Bar: True, [1, 2, 3]>

Any fields inside the type must either be Buildable or one of the following types:

a function

a tuple (up to 8-tuples)

list, NonEmpty, Seq

Map, IntMap, Set, IntSet

Maybe, Either

The exact format of genericF might change in future versions, so don't rely on it. It's merely a convenience function.

newtype GenericBuildable a

A newtype for deriving a generic Buildable instance for any type using DerivingVia.

>>> :set -XDerivingVia
>>> :{
data Bar = Bar { x :: Bool, y :: [Int] }
  deriving stock Generic
  deriving Buildable via GenericBuildable Bar
:}

>>> pretty (Bar True [1,2,3])
Bar:
  x: True
  y: [1, 2, 3]

Constructors

GenericBuildable a

class GBuildable f where

Methods

gbuild :: f a -> Builder

class Buildable' a where

A more powerful Buildable used for genericF. Can build functions, tuples, lists, maps, etc., as well as combinations thereof.

Methods

build' :: a -> Builder

class GetFields f where

Methods

getFields :: f a -> [(String, Builder)]

Get fields, together with their names if available